requirements.txt
bcrypt
deepface==0.0.75
dill
fastapi
pandas
passlib
Pillow
pymongo
python-dotenv
python-jose
python-multipart
PyYAML
uvicorn
-e .